Move Component definition outside of api package
